refactor: Round 7 codebase health - extract ghost, ai, player modules
Split 3 files (2027 LOC) into 8 focused modules (1348 LOC, 33% reduction):

- ghosts/manager.py: 710 → 249 LOC
  - ghost_spawning.py: Ghost placement and Echo paths
  - ghost_behaviors.py: Per-turn behavior processing
  - ghost_trials.py: Champion trials and Hollowed spawning

- ai_scoring.py: 700 → 531 LOC
  - ai_scoring_hazards.py: Hazard pathfinding and pressure analysis

- player.py: 617 → 568 LOC
  - player_feats.py: Feat bonus aggregators
